moved the process of uploading image to image service
