+++ Annotations.API/Groups/ImageEndpoints.cs
@@ -53,7 +53,7 @@ public static class ImageEndpoints
     //and then gets uploaded into your local Azurite BlobStorage as a JSON file
 
     public static async Task<IResult> UploadImageHandler(IFormFile image, string category, AnnotationsDbContext context, 
-        [FromServices] IAzureClientFactory<BlobServiceClient> clientFactory, [FromServices] IImageService _imageService)
+         [FromServices] IImageService _imageService)
     {
         ValidationResponse response = _imageService.ValidateImage(image);
         if (!response.Success)
@@ -65,63 +65,9 @@ public static class ImageEndpoints
         }
         else
         {
-            var containerClient = _imageService.createContainer(clientFactory);
-
-
-            //fileExtension will always be a proper fileExtension because of the ValidateImage method
-
-
-            using (MemoryStream ms = new MemoryStream())
-            {
-                await image.OpenReadStream().CopyToAsync(ms);
-                var thisImage =
-                    new ImageModel() //TODO: dont do this - the title, description and datasetsId are hardcoded
-                    {
-                        Id = _counter,
-                        Title = "idk",
-                        Description = "description",
-                        ImageData = ms.ToArray(),
-                        Category = category,
-                        DatasetsIds = new List<int>() { 1, 2 },
-                    };
-                /*
-                Below functionality of cross-adding the image to the assigned datasets
-                is overriden by the hard-code creation of datasets in Api, Program.cs
-                The idea for future reference is that when uploading an image, the user
-                inputs the names (ids) of datasets, and we can use below code (without
-                hard-coded ids) to cross-add the images to the relevant datasets.
-                */
-                var neededDataset = context.Datasets.Select(Dataset => Dataset)
-                    .Where(Dataset =>
-                        Dataset.Id == 1 || Dataset.Id == 2); //TODO dont do this - this is hardcoded for testing
-
-                foreach (Dataset dataset in neededDataset)
-
-                {
-                    dataset.ImageIds.Add(thisImage.Id); //adds images to the datasets
-                    await context.SaveChangesAsync();
-                }
-
-                string jsonString = System.Text.Json.JsonSerializer.Serialize(thisImage); //objects becomes JSON string
-                var byteContent = System.Text.Encoding.UTF8.GetBytes(jsonString); //JSON string becomes byte array
-
-
-                BlobClient thisImageBlobClient = containerClient.GetBlobClient($"{_counter}.json");
-                _counter++;
-
-                var blobHeaders = new BlobHttpHeaders
-                {
-                    ContentType = "application/json"
-                };
-
-                // Trigger the upload function to push the data to blob
-                await thisImageBlobClient.UploadAsync(new MemoryStream(byteContent),
-                    blobHeaders); //uploaded as byte array
-                //Should it be uploaded as a string instead? So far all endpoints retrieve the JSON files as strings?
-            }
-
-
-
+            
+            await _imageService.UploadingImage(image,_counter, category);
+            _counter++;
             return Results.StatusCode(200);
         }
     }
@@ -130,7 +76,7 @@ public static class ImageEndpoints
     {
         var cts = new CancellationTokenSource(5000);
 
-        var containerClient = _imageService.createContainer(clientFactory);
+        var containerClient = _imageService.createContainer();
         BlobClient blobClient = containerClient.GetBlobClient(imageId + ".json");
         if (!blobClient.Exists(cts.Token).ToString().Contains("404"))
         {
@@ -153,7 +99,7 @@ public static class ImageEndpoints
 
         //enters images
 
-        var containerClient = _imageService.createContainer(clientFactory);
+        var containerClient = _imageService.createContainer();
         BlobClient blobClient = containerClient.GetBlobClient(imageId + ".json");
         if (!blobClient.Exists(cts.Token).ToString()
                 .Contains("404")) //checks if the blobClient is empty/couldn't find the image of that format
@@ -175,7 +121,7 @@ public static class ImageEndpoints
     {
         var cts = new CancellationTokenSource(5000);
 
-        var containerClient = _imageService.createContainer(clientFactory);
+        var containerClient = _imageService.createContainer();
 
         var listOfFiles = containerClient.GetBlobsAsync().AsPages();//all data inside of blobContainer
               
@@ -256,7 +202,7 @@ public static class ImageEndpoints
             };
             
             HashSet<string> collection = new HashSet<string>();
-            var containerClient = _imageService.createContainer(clientFactory);
+            var containerClient = _imageService.createContainer();
 
             //foreach (DatasetModel datasetModel in datasets)//guaranteed to only be one dataset in "datasets", so this is not linear time. 
             //{//there is a better way of doing this
